Refactor error checks in to separate module and move code in to src folder Fixes #54
